refactor: QuackTextStyle.times 로직 개선

+++ ui-components/src/main/kotlin/team/duckie/quackquack/ui/typography/QuackTextStyle.kt
@@ -164,7 +164,7 @@ class QuackTextStyle internal constructor(
 
 @Stable
 private operator fun TextUnit.times(times: Long) = TextUnit(
-    value = value * times,
+    value = (value.toDouble() * times.toDouble()).toFloat(),
     type = type,
 )
